OKX/requirements.txt
# OKX 量化交易机器人依赖包

# HTTP请求
requests>=2.31.0

# WebSocket
websocket-client>=1.6.0

# 配置文件解析
PyYAML>=6.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# 日志
colorlog>=6.7.0

# Web UI
flask>=3.0.0
flask-cors>=4.0.0

# 绘图（回测）
matplotlib>=3.7.0
